backend/pyproject.toml
[project]
name = "lionweather-backend"
version = "1.0.0"
description = "LionWeather API - AI-powered weather forecasting"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "statsmodels>=0.14.0",
    "prophet>=1.1.5",
    "tensorflow>=2.15.0",
    "hypothesis>=6.92.0",
    "apscheduler>=3.10.0",
    "aiohttp>=3.9.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
